--- src/components/MarkdownEditor-mdx.tsx
@@ -1,28 +1,0 @@
-import React, { useState } from 'react';
-import { KitchenSinkToolbar, MDXEditor, codeBlockPlugin, codeMirrorPlugin, diffSourcePlugin, frontmatterPlugin, headingsPlugin, imagePlugin, linkDialogPlugin, linkPlugin, listsPlugin, markdownShortcutPlugin, quotePlugin, tablePlugin, thematicBreakPlugin, toolbarPlugin } from '@mdxeditor/editor';
-
-export const ALL_PLUGINS = [
-  //toolbarPlugin({ toolbarContents: () => <KitchenSinkToolbar /> }), //Gets error about missing plugin or label viewMode
-  listsPlugin(),
-  quotePlugin(),
-  headingsPlugin(),
-  linkPlugin(),
-  linkDialogPlugin(),
-  imagePlugin(),
-  tablePlugin(),
-  thematicBreakPlugin(),
-  frontmatterPlugin(),
-  codeBlockPlugin({ defaultCodeBlockLanguage: 'txt' }),
-  codeMirrorPlugin({ codeBlockLanguages: { js: 'JavaScript', css: 'CSS', txt: 'text', tsx: 'TypeScript' } }),
-  diffSourcePlugin({ viewMode: 'rich-text', diffMarkdown: 'boo' }),
-  markdownShortcutPlugin()
-]
-
-const MarkdownEditor: React.FC = () => {  return (
-    <div>
-      <MDXEditor onChange={console.log} markdown='type here' plugins={ALL_PLUGINS} />
-    </div>
-  );
-};
-
-export default MarkdownEditor;
